requirements.txt
pip==19.0.3
Flask==1.0.2
numpy==1.15.4
numexpr==2.6.9
matplotlib==3.0.2
google-cloud-storage
